refactor ReviewSession into multiple components
